[pt1][quant] Add dynamic quantized Linear op in PyTorch
As suggested in https://github.com/pytorch/pytorch/pull/22891, we will add an overload for torch.fbgemm_linear_int8_weight (dynamic quantized version of linear function) that takes PackedLinearWeight as input and is pretty much the same in signature as regular aten::linear.

The previous Diff D16381552 is reverted because `quantize_linear` expects the scale to be `float`, and the zero_point to be `int`.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/23464
ghstack-source-id: 87539311

Differential Revision: [D16527741](https://our.internmc.facebook.com/intern/diff/D16527741/)