[jit] serialize all c++ frontend modules to a single CU.
Previously, every module would get its own CompilationUnit when saving
from the C++ frontend. That's bad because nothing checks that they have
uniquely qualified names or mangles them to make them unique.

This was okay when we were doing model.json, but once we serialize
modules like classes this will cause an error on import (when we try to
re-define the same class a bunch of times.

ghstack-source-id: 2f29a530e0adc9274f05b8495f1117efe0f1da23
Pull Request resolved: https://github.com/pytorch/pytorch/pull/23645